requirements.txt
nats-py==2.1.0
cryptography==3.4.8
PyYAML==6.0
click==8.0.3
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-mock==3.7.0